+++ spec/rubocop/cop/capybara/specific_finders_spec.rb
@@ -12,6 +12,29 @@ RSpec.describe RuboCop::Cop::Capybara::SpecificFinders do
     RUBY
   end
 
+  it 'registers an offense when using `find` with kind' do
+    expect_offense(<<~RUBY)
+      find(:css, '#some-id')
+      ^^^^^^^^^^^^^^^^^^^^^^ Prefer `find_by_id` over `find`.
+    RUBY
+
+    expect_correction(<<~RUBY)
+      find_by_id('some-id')
+    RUBY
+  end
+
+  it 'registers an offense when using `find` with kind ' \
+     'and option' do
+    expect_offense(<<~RUBY)
+      find(:css, '#some-id', class: 'some-cls')
+      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Prefer `find_by_id` over `find`.
+    RUBY
+
+    expect_correction(<<~RUBY)
+      find_by_id('some-id', class: 'some-cls')
+    RUBY
+  end
+
   it 'registers an offense when using `find` with no parentheses' do
     expect_offense(<<~RUBY)
       find "#some-id"
@@ -149,6 +172,30 @@ RSpec.describe RuboCop::Cop::Capybara::SpecificFinders do
 
     expect_correction(<<~RUBY)
       find_by_id('some-id')
+    RUBY
+  end
+
+  it 'registers an offense when using `find` ' \
+     'with argument is kind and attribute specified id' do
+    expect_offense(<<~RUBY)
+      find(:css, '[id=some-id]')
+      ^^^^^^^^^^^^^^^^^^^^^^^^^^ Prefer `find_by_id` over `find`.
+    RUBY
+
+    expect_correction(<<~RUBY)
+      find_by_id('some-id')
+    RUBY
+  end
+
+  it 'registers an offense when using `find` ' \
+     'with argument is kind and attribute specified id and option' do
+    expect_offense(<<~RUBY)
+      find(:css, '[id=some-id]', class: 'some-cls')
+      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Prefer `find_by_id` over `find`.
+    RUBY
+
+    expect_correction(<<~RUBY)
+      find_by_id('some-id', class: 'some-cls')
     RUBY
   end